Added: Create socket using protocol family that is compatbile with the address family of an IPEndPoint. Added: Create copy of IPEndPoint with modified port via ctor. Fixed: Added IPv6 support to ClusterDiscoveryService. THIS RESULTED IN THE FOLLOWING BREAKING CHANGES:   - The user message size has been reduced to 128 Byte.   - The magic number and message type identifier field in the datagram have been increased to 32 bit to retain a consistent alignment on all supported platforms   - The magic number has been changed to reflect the breaking change Fixed: Shutdown of cluster discovery listener is now implemented via closing the communication channel. Therefore, the listener thread is unblocked on Terminate() in any case. Fixed: The native interoperability class for ClusterDiscoveryListener has been updated to reflect the changes above.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@647 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ net/include/vislib/ClusterDiscoveryListener.h
@@ -101,7 +101,7 @@ namespace net {
          */
         virtual void OnUserMessage(const ClusterDiscoveryService& src,
             const ClusterDiscoveryService::PeerHandle& hPeer, 
-            const UINT16 msgType, const BYTE *msgBody);
+            const UINT32 msgType, const BYTE *msgBody);
     };
     
 } /* end namespace net */
